refactor: Separate thunks
