+++ src/lib/constants.js
@@ -1,0 +1,3 @@
+export const baseUrl = "http://webshopdemo.devweb.b-s.si";
+export const tenant = "WebShopDemo";
+export const company = "FLB";
