+++ src/pages/Orders.js
@@ -1,7 +1,6 @@
 import emptyOrder from "../assets/empty.1024x801.png";
 import EmptyContent from "../components/EmptyContent";
 import { useDispatch, useSelector } from "react-redux";
-import { fetchOrderItemsForOrder, fetchOrders } from "../redux/orderSlice";
 import { useEffect, useState } from "react";
 import { useAuth } from "../context/authContext";
 import BasketTable from "../components/BasketTable";
@@ -13,6 +12,7 @@ import {
   ListItemText,
   Typography,
 } from "@mui/material";
+import { fetchOrderItemsForOrder, fetchOrders } from "../api/ordersThunks";
 
 const Orders = () => {
   const { authToken } = useAuth();
